+++ djangae/contrib/search/apps.py
@@ -1,6 +1,5 @@
 
 from django.apps import AppConfig
-from django.core.exceptions import ImproperlyConfigured
 
 
 class SearchConfig(AppConfig):
@@ -8,9 +7,4 @@ class SearchConfig(AppConfig):
     verbose_name = "Search"
 
     def ready(self):
-        try:
-            import nltk  # noqa
-        except ImportError:
-            raise ImproperlyConfigured(
-                "djangae.contrib.search depends on the nltk library, which is not installed"
-            )
+        pass
