+++ djangae/contrib/search/document.py
@@ -9,6 +9,37 @@ PUNCTUATION = {
     "}", "~", "$"
 }
 
+# These are the English stopwords generated by nltk. Ideally
+# we'd use nltk directly, but that involves some mechanism of
+# downloading/shipping the nltk data files. That's something
+# for the future.
+STOP_WORDS = [
+    'i', 'me', 'my', 'myself', 'we', 'our', 'ours', 'ourselves',
+    'you', "you're", "you've", "you'll", "you'd", 'your', 'yours',
+    'yourself', 'yourselves', 'he', 'him', 'his', 'himself', 'she',
+    "she's", 'her', 'hers', 'herself', 'it', "it's", 'its', 'itself',
+    'they', 'them', 'their', 'theirs', 'themselves', 'what', 'which',
+    'who', 'whom', 'this', 'that', "that'll", 'these', 'those', 'am',
+    'is', 'are', 'was', 'were', 'be', 'been', 'being', 'have', 'has',
+    'had', 'having', 'do', 'does', 'did', 'doing', 'a', 'an', 'the',
+    'and', 'but', 'if', 'or', 'because', 'as', 'until', 'while', 'of',
+    'at', 'by', 'for', 'with', 'about', 'against', 'between', 'into',
+    'through', 'during', 'before', 'after', 'above', 'below', 'to',
+    'from', 'up', 'down', 'in', 'out', 'on', 'off', 'over', 'under',
+    'again', 'further', 'then', 'once', 'here', 'there', 'when',
+    'where', 'why', 'how', 'all', 'any', 'both', 'each', 'few',
+    'more', 'most', 'other', 'some', 'such', 'no', 'nor', 'not',
+    'only', 'own', 'same', 'so', 'than', 'too', 'very', 's', 't',
+    'can', 'will', 'just', 'don', "don't", 'should', "should've",
+    'now', 'd', 'll', 'm', 'o', 're', 've', 'y', 'ain', 'aren',
+    "aren't", 'couldn', "couldn't", 'didn', "didn't", 'doesn',
+    "doesn't", 'hadn', "hadn't", 'hasn', "hasn't", 'haven',
+    "haven't", 'isn', "isn't", 'ma', 'mightn', "mightn't", 'mustn',
+    "mustn't", 'needn', "needn't", 'shan', "shan't", 'shouldn',
+    "shouldn't", 'wasn', "wasn't", 'weren', "weren't", 'won', "won't",
+    'wouldn', "wouldn't"
+]
+
 
 class Field(object):
     def __init__(self, default=None, null=True):
@@ -16,10 +47,6 @@ class Field(object):
         self.null = null
 
     def normalize_value(self, value):
-        from nltk.corpus import stopwords  # Inline import so not a hard requirement
-
-        stop_words = stopwords('english')  # FIXME: Multiple languages
-
         # Default behaviour is to lower-case, remove punctuation
         # and then remove stop words
 
@@ -28,7 +55,7 @@ class Field(object):
 
         value = value.lower()
 
-        to_remove = set(stop_words).union(PUNCTUATION)
+        to_remove = set(STOP_WORDS).union(PUNCTUATION)
 
         tokens = value.split()  # Split on whitespace
         tokens = [x for x in tokens if x not in to_remove]
@@ -75,7 +102,10 @@ class Document(object):
 
                 # Apply any field values passed into the init
                 if attr in kwargs:
-                    setattr(self, attr, kwargs[attr])
+                    setattr(self, attr_name, kwargs[attr])
+                else:
+                    # Set default if there was no value
+                    setattr(self, attr_name, attr.default)
 
     @property
     def id(self):
